Add and begin using printTokens debug utility
The conversion away from pyatspi means we don't get the stringifiers
for accessible objects, events, etc. Thus our existing debugging
strings are far less helpful. See issue #300.

In addition, as was pointed out in issue #389, including an accessible
object in a debugging string invokes pyatspi's stringifier which
includes retrieving its name. This work happens before we check if
debugging is enabled. Switching to printTokens, which returns early
if debugging is not enabled, will eliminate that work.