Fix delay when a key is held down
For historical reasons, we currently stuff an incoming event into an
AtspiDeviceEvent, but the new AtspiDevice API doesn't include a timestamp
for events, and AtspiDeviceEvent defines the timestamp as a guint. We were
stuffing the value of time.time() into this timestamp value, which cast it
to an int and made it accurate only to the second, making further events have
the same timestamp as the original and appear to be duplicates.

Fixes #390