Don't cut off speech if focus changes away from checkable item
Now that we interrupt speech due to focus changes (due to the
potential of not getting input events in Gtk4 + Wayland), there's
the possibility we'll cut off important information. Here's yet
another scenario:

The user presses return on a check menu item or radio menu item
causing the checked/selected state to change *and* the menu to
close returning focus to a push button or something else. We
don't want that focus change to cut off the "checked/selected"
announcement.