Don't treat prioritized events as duplicates during flood
We normally want to ignore duplicate events. However, in MATE we're
seeing a prioritized event (e.g. focused:true), followed by an event
flood, followed by a duplicate of that prioritized event. If we
ignore the duplicate, our logic to prune events during floods gets
bypassed. Therefore if we are in a flood, do not treat prioritized
events as duplicates even when they technically are.