Create a utility to compare strings for redundancy
We're seeing cases where the label and/or name and/or description
are essentially the same with only slight differences in wording.
This makes Orca chatty. Create a function we can use to compare
strings and eliminate some of this chattiness.

Note that this commit doesn't make use of the new function. That
will be done incrementally so that we can adjust or undo things
cleanly.