Set default max row and col count for isNonUniformTable
If a table is giant, checking each cell on the off chance that it
spans more than one row or column so we can announce it as "non-
uniform" is not worth the performance hit. Check at most 25 columns
within 25 rows by default. If it's still "uniform" at that point,
it will probably continue to be.

See issue #398.