Work around more Qt brokenness
8f1b2c58abdd3bab89a4eb7caeb25371b6ed1875 regressed flat review of
some Qt apps (at least qjoypad) because it expected there to be an
object with a frame/window role and/or a dialog role containing the
focused object. Unfortunately, some Qt apps don't expose valid roles
for the top-level object. The top-level object should never have a
role of filler, for instance. In order to not regress the user
experience so close to the release, we'll work around this brokenness
for now by accepting whatever ancestor descends from the application
role in Qt apps as the top-level object.