Add experimental setting for flat review restriction
Issue #380 was filed requesting we save the flat review restriction
setting. We deliberately do not permanently save settings which have
been changed on the fly. Things that should be saved permanently
should have a corresponding item in Orca's preferences dialog.

That said, some local experimentation with persisting the setting
revealed some possibly unwanted side effects. For instance, it is
really easy to get trapped in something small (e.g. a list/tree
item that just happened to be focused when entering flat review).
And there's no indication that this has happened; it just seems
like flat review stopped working.

Some questions to be answered include:

1. Is persisting it, even via GUI, a good idea at all?
2. Should restriction be limited to true/false (current), or
   should we add additional logic to identify the most suitable
   container? For instance, if the current focus is a list/tree
   item, restricting review to the list/tree makes a lot more
   sense than restricting it to the item.
3. Should there be an announcement that we make when flat
   review has been entered in restricted mode, at least at the
   verbose verbosity level?

In addition, if we add a GUI option that makes it possible to
restrict flat review, we also need to provide a default binding.
Otherwise, we're making it possible for users to get stuck and
no ready-to-use command to get unstuck.

To make it easier to start answering these questions, create an
official setting for flat review restriction. In order to make it
persist, it must be added to $HOME/.local/share/orca_customizations.py:

orca.settings.flatReviewIsRestricted = True