Flat Review: Always start at the object of interest
When you enter flat review, Orca should always start at the object of
interest. In focus tracking mode, that object is the locusOfFocus;
in the object navigator, that object should be the current navigator
object. This was not occurring unless flat review restriction was
enabled, which was done as part of the new restriction option. These
two things need to be decoupled.

Flat-review restriction has nothing to do with the object navigator;
if restriction is enabled, it means that the only contents that are
reviewable are the current object and any descendants of that object.
If restriction is disabled, the entire window is reviewable.

Because of the incorrect association with restriction and the object
of interest, entering flat review from the object navigator resulted
in flat review starting at the locus of focus and not the navigator
object. Fix that by always starting flat review at the object of
interest.