Be less enthusiastic about declaring objects permanently "dead"
Accessible objects go defunct on us frequently. For performance, it is
best to identify such "dead" objects as soon as possible and stop doing
any work related to an object known to be dead. In Orca v45, the logic
related to dead-object identification was modified to cache objects
known to be dead (according to the exceptions we got from AT-SPI2).
However, it seems like objects are being reported with a "does not
exist" exception, and later becoming valid. In reduce (and hopefully
eliminate) false positives:

* Keep track of both known-dead and known-alive objects
* Only mark something as known-dead if its current status is known-alive.
  This should make it possible to distinguish objects we care about that
  are now dead, from objects we get events for that are not yet fully
  alive.
* Set/update the status back to alive if a common Atspi call (to get the
  name, role, or state) succeeds without exception.