Use AT-SPI2's cache for everything
Now that AT-SPI2 provides a way for us to clear the cache of a single
accessible object (as opposed to the cache of the entire subtree), we
can resume using its cache for everthing. We were not caching the name
and description because we didn't want to present stale information,
but we also wanted to avoid the potential performance hit from clearing
the cache of a potentially large subtree. But not using the cache for
name and description comes with its own (albeit smaller) cost. Now we
can avoid that by clearing the cache non-recursively in setLocusOfFocus.
And because we do this before presenting the object, we should have
the most up-to-date info about that object.