Action Presenter: Show menu after brief timeout
In the new-keyhandling world, GtkMenu is almost always destroying our
menu because it failed to make a keyboard grab. As suggested in issue
around this problem. And at least locally, one tenth of a second seems
to be sufficient. So start with that instead of a full second. A follow-up
commit will attempt to further improve the user experience by processing
windows which hold a single menu more quickly.