Focus Manager: Remove desktop-frame check from can_be_active_window
We have historically had to check (and eliminate) the GNOME2/MATE
desktop frame as the active window because it claimed to be active
when it wasn't. Prior to the creation of the focus manager, this
check was only done if we had multiple windows claiming to be active.
Moving it to can_be_active_window caused us to not present a Caja
window:activate event when the desktop gained focus.

Fix this regression by removing the check from can_be_active_window.
Do not restore it to the sanity-check code. If we discover Caja is
not updating its state, that bug should be fixed in Caja rather than
worked around in Orca.