Merge branch 'brlapi_priority' into 'master'
Set BRLAPI priority to 70 when using flat review to avoid it to be
locked by BRLTTY.

Since BRLTTY manages terminal access, it is not possible to use flat
review with Braille. To allow it again, Orca has to increase BRLAPI
priority to 70 to be higher than BRLTTY (60). This MR does the
following:

* Implements `braille.setBrlapiPriority()` with a parameter `level`
  with a default value of 50
* Adds a call to `focus_manager.emit_region_changed()` with `level`
  set to 70 if the new mode is flat review, or to the default value
  otherwise.

See merge request GNOME/orca!177