Don't use cache when retrieving the heading level for an editable object
We now use cached object attributes globally for performance. There are
cases in which we deliberately don't because we suspect it is likely
for the value to change. If the heading level changes in an editable
document, we don't want to present stale information. Therefore, only
used the cached value for heading level on non-editable objects.

Fixes issue #423.