Check that cell is a really a cell when getting row/column headers
It turns out that the braille generator is globally asking for the
row and column headers when generating ancestors, regardless of
whether we're in a table descendant. That needs to be changed. In
the meantime, return immediately if we're seeking row/column headers
for anything other than a table cell.