Remove redundant call to UpdateBraille in table cell presentation
The structural navigation code now uses the generators, rather than
custom presentation, for table cells. This causes us to present the
cell in braille, making the call to UpdateBraille redundant. Also,
UpdateBraille has some logic which updates the caret context. That
can cause us to believe we were already in a particular cell and
fail to speak the header because it's not "new."