Add one more sanity check and assume the row/colspan of tree cells is 1
It's assumed that a cell in a tree (which arguably should have the
treeitem role) will not span multiple rows or columns. So we can save
some work here. In addition, it appears that Qt has trees that implement
the table interface, have child cells, and segfault upon calls to
Atspi.Table.get_row_column_extents_at_index() due to a "bad" or unexpected
index. That will be debugged and filed next. In the meantime, we don't
want the much-needed table rewrite to regress the user experience by
crashing Qt apps....