Remove the GAIL script; use gtk script instead
The GAIL script was for Gtk2. We had created and maintained a separate
gtk script for Gtk3 thinking things might be(come) radically different.
In the end the scripts largely wound up with quite similar logic. Rather
than maintain a dedicated, redundant script for whatever Gtk 2 apps may
still be in use, we can use the gtk script and update it to handle any
Gtk2isms should they appear.