Gtk: Ignore focus change events for ancestor of focused item
Given a tree table in which there is a focused item that is not
selected, we sometimes get an active-descendant changed event for
the focused item followed by a focused event for the tree table
that contains the item. Should this occur, ignore the focused event
as it is redundant to the focused item we just presented.