Fix regression in pidgin child node presentation
As part of the table refactor, if the row index or the column index
was -1, we treated it as an error condition and returned -1 for both.
This regressed the presentation of the child nodes in pidgin due to
the fact that the information we need is in a hidden cell with a
corresponding column index of -1. Fix that by returning the indices
as-is. Scripts will have to do any error handling, but they should
already be checking for valid values.

Also restore incrementing the column index in childNodes. That
adjustment was removed accidentally.