Add means to suspend structural and/or caret navigation support
Structural navigation and caret navigation can be toggled on/off
independently of focus mode. The new keyhandling does not handle
this. This commit adds a means to suspend support without disabling
it in focus mode. Note that nothing is using this functionality yet
and there should not be any change in existing behavior.