Suspend caret navigation and structural navigation in focus mode
This is needed to remove the input event handlers and keybindings
from the web script and is a precursor to eliminating the web
script's getEnabledKeyBindings override. Making that change will
be done as a dedicated commit in case it needs to be reverted while
we sort out more new-keyhandling bugs.