Eliminate the default script's getEnabledKeyBindings
This seems to have been added to deal with some web-script-specific
scenarios that (hopefully) no longer exist. Therefore move the logic
to addKeyGrabs. In addition, clear the existing key bindings in
removeKeyGrabs. It appears that we were removing the grabs but
leaving the keybindings for the script as-is. Therefore any caller
which was accessing the script's bindings directly (which should not
be done) would have out-dated information.