Reland the two changes related to getEnabledKeyBindings
The source of the thunderbird regression -- and the subsequently
reported regression about toggling browse/focus mode failing --
is that refreshKeyGrabs was getting called without up-to-date
information about whether or not caret and structrual navigation
should be used. Fix that by always checking whether or not we
should suspend this navigation and call suspend_navigation prior
to refreshKeyGrabs.

A subsequent update will further clean this logic up once we've
verified there are no additional regressions.