overrideKeyBindings should not add bindings for removed handlers
overrideKeyBindings was unconditionally adding user bindings to the
script, even when the command had been deliberately removed. As a result,
we were seeing instances where structural navigation bindings were
correctly removed by structural navigation in response to entering
focus mode, only to be re-added by the default script as a side effect
of having been rebound to a different shortcut by the user.

Modify overrideKeyBindings to only update keys associated with
handlers that exist in the script.