Fix regression in preferences dialog related to new keyhandling
The caret and structural navigation bindings went missing from the
preferences dialog when in focus mode. In addition, the "modified"
checkbox was no longer checked for modified bindings. Both of these
were due to removing "suspended" bindings and input event handlers
from the script. For instance, in focus mode the structural navigation
and caret navigation keys are suspended.

In order to fix this:
* add an "enabled" property to InputEventHandler and KeyBinding
* set "enabled" to false when the associated command is suspended
  or not enabled
* do not remove any bindings or handlers from the script
* only add key grabs for bound bindings which are enabled
* modify overrideKeyBindings so that we can ask for all keybindings
  rather than the default enabled keybindings. This ensures we only
  apply user overrides when they apply (the functionality is active
  or we're in the preferences dialog).

See issue #428.