Web: Do not remove key grabs in response to window deactivation
For some reason the new keyhandling landed with a call to remove
key grabs in response to window deactivation. All other scripts
(including the web script) remove key grabs when the script itself
becomes inactive and adds key grabs when the script itself becomes
active. It's not clear to me what circumstance motivated the additional
removal, but here's why it's a bad idea: Web apps with dialogs can
result in a window deactivation event when the dialog is dismissed.
But the web script will remain active, so we remove all key grabs and
no Orca commands whatsoever work in the browser until that window is
left and returned to. Fix that by removing the call to removeKeyGrabs
in onWindowDeactivated.