Add means to identify and quickly dismiss obsolete events
Given an event flood, or rapid navigation, we can wind up with a
huge queue of events. Add a check which determines if a more recent
event obsoletes an event we are about to process. This is a work in
progress, but seems to speed up fast arrowing in widgets and text
quite a bit.