Restore workaround for broken tables for flat review
As part of the complete rewrite of Orca's table support, old hacks
which no longer seem needed were removed. One of those hacks was to
manually retrieve and add the column headers to the flat review context
for tables which did not allow retrieval via Table.get_accessible_at.
That hack appears to still be needed for (at least) Thunderbird 110.
It is not needed for Thunderbird 115. Put the hack back for now so
as not to regress the user experience.