Remove the old and outdated tutorial generator and its messages
As discussed on the mailing list in September[1], the existing tutorial
messages have many problems, one of which is that they assume shortcuts
based on a GNOME 2.x + GTK world. That was a long time ago. Things have
changed, more toolkits have become accessible, and hard-coded strings
that may not even be accurate can produce more harm than good. This
commit is the first step towards resolving that: The tutorial generator
and its messages have been removed.

In its place is functionality to retrieve "help text." Orca asks for the
help text via Atspi.Accessible.get_help_text() and falls back on the
"helptext" object attribute. The former does not exist yet, but I'm
hopeful it will. The latter is intended as a means through which
application and toolkit developers can work on a prototype and give
feedback on Orca's presentation without having to wait for the Atspi API
requested in https://gitlab.gnome.org/GNOME/at-spi2-core/-/issues/146.

Note that most users won't be impacted by this change because tutorial
messages are disabled by default.

See issue #432.

[1] https://www.freelists.org/post/orca/Thoughts-about-tutorial-messages