KeyboardEvent: Reset click count if object changes after handling
If a user presses and releases a key multiple times in quick succession,
we treat that as a double or triple click without much checking. That is
fine in the case of commands that don't update the locusOfFocus. But if
a user re-maps one navigation command to be a double-click of a different
navigation command, we can mistakenly treat two back-to-back instances of
command 1 as a single instance of command 2.

One way to reduce the likelihood of such a false positive is to see if
the user's location changed as a result of our handling the previous
event. If it has, do not treat the incoming event as a double/triple
click.