New keyhandling: Eliminate web-related calls to refreshKeyGrabs
The refreshKeyGrabs function landed with a TODO stating that it probably
should not do what it was doing. That TODO was accurate, especially in
the case of the web script where modes can be rapidly and frequently
toggled and where there are a lot (over 200) bindings and grabs, many
of which do not need to be refreshed.

In order to eliminate all web-related calls to refreshKeyGrabs:
* Create refresh_bindings_and_grabs for each module where frequent
  changes are anticipated. This function's job is to update its own
  bindings, grabs, etc.
* Ensure this function gets called any time the associated module's
  supsended or enabled state changes.
* Remove the no-longer-needed calls to refreshKeyGrabs from the web
  script.

This change should improve responsiveness during mode changes and
greatly reduce (hopefully eliminate) the command leak-through cited
in the original TODO.

At this time, the only remaining caller of refreshKeyGrabs is the
preferences dialog when keys are being rebound.