Create a globally available Table Navigator
Orca's existing cell navigation should now work everywhere, not
in document content. The unchanged commands are:
* Alt+Shift+Left/Right/Up/Down: move to the next cell in the
  specified direction
* Alt+Shift+Home/End: move to the first/last cell of the table.

The following commands have been added:
* Orca+Shift+T: toggle table navigation (it's enabled by default)
* Orca+Alt+Shift+Left/Right/Up/Down: move to the final cell in
  the specified direction.

NOTE: Orca+Z no longer applies to table cell navigation. But in
the case of web content, table cell navigation should automatically
be toggled off/on for focus mode and browse mode as before. If that
does not happen, it is a bug.

Orca's dynamic header commands from LO should now work everywhere,
including Google Docs and other web-based editors.

NOTE: These bindings have changed. The new bindings are:
* Orca+Shift+R: Set the cells to be treated as column headers to
  this row.
* Orca+Shift+R double-click: Unset the dynamic column headers row.
* Orca+Shift+C: Set the cells to be treated as row headers to this
  column.
* Orca+Shift+C double-click: Unset the dynamic row headers row.

This change was done to free up Orca+R for a future feature. But
given the need for a change, the opportunity was taken to use the
same bindings as NVDA.