Ensure we remove bindings and grabs for overridden bindings
2612b787753176 replaced our use of the heavy refreshKeyGrabs with
a much lighter change-only-what-needs-changing approach. And thus
regressions have presented themselves.

This commit fixes one of the two regressions, namely the original
binding continuing to do its original function even though it was
overridden. In the old, heavy way all bindings were refetched and
overrideKeyBindings removed the old and enabled the new. We need
to do that work in the individual refresh_bindings_and_grabs.

As part of some related cleanup, eliminate removeByHandler and use
remove instead. The fewer places we're messing with bindings and
grabs, the better.

The remaining (known) regression is that we are still grabbing
the user's chosen override. A fix for that is still pending.