Address regression in which grabs are not removed for overridden bindings
2612b787753176, which replaced our use of the heavy refreshKeyGrabs with
a much lighter approach to only change what needs to be changed revealed
a bug in which grabs were getting added but never removed. The old logic
removed every last keygrab (200+ in some cases) and then recreated them
all from scratch. Something in that thoroughness was removing the grab
in question.

Properly debugging and fixing the underlying issue is still pending. In
the meantime, catch and handle that condition in Keybindings.remove()
so that we don't have to revert the performance win in order for users
to test the new table navigator.