Fix keyhandling focus-mode regression
overrideKeyBindings was created years ago and with the assumption
that we'd be working with a full set of script handlers and script
bindings. In the new-keyhandling, only-change-what-we-need world,
this assumption is no longer valid. All callers have and know are
their own handlers and their own bindings. Therefore, modify
overrideKeyBindings to take a set of handlers rather than a script.

See issue #435.