Ensure key grabs are refreshed when Chromium-based menus open
e80fd186a6678b6 removed a call to removeKeyGrabs in the web script's
onWindowDeactivated because it interfered with web-app dialogs and it
was not clear why removing key grabs was needed in that case. This
removal regressed mode change for Chromium-based menus because the
Chromium script was updating the focus to the menu in onShowingChanged,
but doing so silently, preventing a call to locusOfFocusChanged. In the
new-keyhandling world, scripts need to be notified of focus changes that
might impact mode changes.

While the logical fix would be to notify the script in onShowingChanged,
it turns out the logic in question is obsolete: it landed five years ago
when Chromium accessibility support was still early in its development
and was failing to fire the expected events for popup menus. Because we
now get the expected window-activate events from Chromium, we can remove
this special case. Doing so fixes the key grab regression because the
script is notified of the focus change.