Ensure locusOfFocusChanged gets called for Thunderbird popup menus
In the new-keyhandling world, any focus change which might require
a corresponding mode change needs to call locusOfFocusChanged.
The Thunderbird script's onShowingChanged was not doing so. While I
am not aware of any keygrab-related bugs from this failure, one could
theoretically exist, so have the focus manager notify the script just
in case.