Remove the old and presumably unused self-voicing script
Nothing in Orca was using it and overriding this via orca-customizations
would take some doing. In addition, there are so many things this script
wasn't handling that it wouldn't really be self-voicing; the default
script would still handle most of the events and present them. Lastly, a
new sleep-mode script will replace the functionality the self-voicing
script was meant to accomplish.