Create a sleep-mode script
This is the first of several steps needed to have sleep-mode support.
It provides the script. There are not (yet) associated commands for
end users to toggle this mode on and off on the fly. That will be a
subsequent change. In the meantime, in order for users to test the
planned support, they can add the app names Orca should sleep for to
orca-customizations.py.

Example: If you want Orca to "sleep" when using pluma and gtk3-demo,
add these lines to orca-customizations.py:

import orca.settings
orca.settings.sleepmodeApps = ["pluma", "gtk3-demo"]

Note that the name must be an exact match with what the app exposes
as its accessible name. In order to get that information, you can
run `orca -l`. The string to be used follows the pid number.

The expected behavior of the sleep-mode script is the following:

1. Orca commands have no effect, at the present time that means
   all of them. When the command is created which makes it possible
   to toggle sleep mode on the fly, that command will of course be
   enabled even in sleep mode.
2. When a sleep-mode application window is activated, Orca presents
   that sleep mode is enabled for that application and then stops
   speaking and brailling anything.
3. Orca ignores all accessible-events (focus changes, selection
   changes, etc.). Please note that "ignores" is not the same as
   "unregisters for." A badly-behaved app that floods Orca with
   events will still flood Orca with events even in sleep mode.
   That is a related, but separate problem.

See issue #243.