Event Manager: Remove _isDuplicateEvent
_isDuplicateEvent's purpose was to stop us from queueing up events
we already had in the event queue for performance purposes. Now
we have _isObsoletedBy which improves performance, which peeks at
the event queue to determine if a newly-dequeued event is still worth
processing. Reasons for the newly-dequeued event no longer being worth
processing include (but are not limited to) a duplicate event subsequently
coming in. Thus it should be safe to remove _isDuplicateEvent.

In addition, there is a scenario in which keeping both functions in
place breaks the user experience. Consider an app which emits the
following events when "foo" gets focus:

1. object:state-changed:focused:true "foo"
2. object:state-changed:focused:false "foo"
3. object:state-changed:focused:true "foo"
4. object:state-changed:focused:false "foo"
5. object:state-changed:focused:true "foo"

Only event 1 -- or arguably event 5 -- is valid. The other four are
being emitted for who knows why. (Plasma shell bug to be filed.) But
if we stop queueing duplicates, events 3-5 will never be queued. Then
_isObsoletedBy will assume event 1 has been obsoleted by event 2 and
Orca is silent.

While apps should not be spewing events like the above, we need to
handle those that do. By removing _isDuplicateEvent, we'll queue up
all 5 events. Then when we dequeue events 1-4, _isObsoletedBy will
immediately discard them. Event 5 has not been obsoleted by anything
(assuming the user has not moved to the next item) and will thus be
presented by Orca.