Table Navigator: Ensure we always use user bindings
All modules for commands that get (un)loaded on the fly (e.g. due to
a switch between browse and focus mode) need to call overrideKeyBindings
to pull in the user's overrides. We were doing this for live regions,
caret navigation, and structural navigation, but failing to do so for
the new table navigator. This commit fixes that.