Clear the cache recursively for table cells
Investigating https://gitlab.gnome.org/GNOME/nautilus/-/issues/3253
led to the discovery that the issue only happened in the grid view;
not the list view. The reason why is that we already had a workaround
for apps and toolkits which fail to fire the events we need. But that
workaround was failing for the grid view because we were only clearing
the cache on the table cell which claimed focus, but looking to the
cell's child for the name/displayed text to present to the user. That
child needs to have its cache cleared as well.

For now, handle this specific case in our workaround. Once we've found
all the apps and toolkits with this issue, and hopefully gotten fixes,
we can stop clearing the cache when setting the locus of focus and
gain some more performance.