Web: Eliminate false positive comparing removed object with focus
When an object is being destroyed, we compare it to the locus of focus
to determine whether or not we need to do some recovery work. Because
the object is being destroyed, we cannot poke at it very much and
expect valid results, therefore we call isSameObject ignoring names
and comparing roles and paths (ignoring the final invalid index in
parent of the object being destroyed). For the most part this works
because paths tend to be quite different. One scenario in which they
are the same is when focus is in a scrollable container which adds
and removes accessible objects as they are scrolled into and out of
view.

In order to eliminate this false positive, only check isSameObject
if the locus of focus is dead. If it's not, and if the equality
check done previously failed, we shouldn't have any need to recover
from the removal of the child.