Remove isActivatableEvent overrides from Gecko and Chromium scripts
The Gecko override is a nearly seven-year-old workaround for bogus
window:activate events from Gecko. Hopefully that bug is fixed. If
it is not, let's get it fixed in Gecko. The Chromium script had the
same override that was copied from the Gecko script when the Chromium
script was first created. There is no indication this override is
actually needed.