Update locus of focus when setting the active window
When we set the active window, it is possible to be in a situation
where the locus of focus is currently not inside the active window.
This rarely happens (e.g. because we set the locus of focus during
window activation, because we immediately get a focus event, etc.).
But any logic based on the current locus of that occurs in between
setting the window and setting focus can be flawed. This appears to
be the reason why browse mode was getting reactivated in VSCode after
switching focus from Firefox in another desktop.

Handle this condition in the focus manager by checking if focus is
inside the current window. If it is not, set focus to the current
window similar to what we do during window activation.