Script manager should handle sleep mode for app "replicants"
Some badly-behaved apps (notably Firefox and Thunderbird) have a
tendency to fire events some events from an application that is not
registered with AT-SPI2 and other events from an application that is.
The script manager does its best to cope with this by finding the
existing script for the "good" app and using it for event processing.

Now that we have sleep mode, we need to check if the "good" app is in
sleep mode. If so, we should use the sleep mode script associated with
the "good" app rather than its default, and inactive, script.

This should fix the issue of Firefox and Thunderbird being in permanent
sleep mode after deactivating and reactivating the window.