Create a system information presenter
This presenter offers quick access to information that is often displayed
in a bar, system tray, widget, etc.

* New command: Present CPU and memory usage (unbound, requires psutil)
* New command: Present battery status (unbound, requires psutil)
* Move date and time presentation into the system information presenter

Note that in both of the new commands, Orca just presents what it gets
via psutil; it does no verification or checking. For example, on the
desktop computer used to develop this feature, psutil reports a battery
whose percentage is always 42%. Issues like this are beyond the scope
of Orca's job to check for and fix.

Also add some error checking for instances where there is a binding
whose handler does not exist. This should never happen, but it is
technically possible (and was encountered due to an error while working
on this feature). Therefore handle it gracefully just in case.