Web: Treat inline iframes as layout only; generate ancestors for descendants
We were not generating ancestors for descendants of inline iframes to
eliminate chattiness with things like facebook "like" buttons. By
side effect, this was causing us to not present ancestor lists when
navigating by line if those lists happened to live inside an iframe
whose reported display style was "inline".

To solve this, treat inline iframes as layout only so that we don't
present them. But do present the ancestry of their descendants.