Ignore selection-changed events from non-showing, non-visible objects
The application prompting this change is VSCode, which is firing
selection-changed on a hidden menu without any name. But in general we
should not be presenting changes and/or updating our state in response
to a selection changes in widgets whose state set suggests they are not
exposed to the user.