Modify shouldInterruptForLocusOfFocusChange to take names into account
If the old locus of focus was an ancestor of the new locus of focus,
we weren't interrupting speech. This ensured no context (like the
newly-activated window) went missing. However, if the old locus of
focus lacks a name, it's likely that presenting its role just adds
chattiness. Therefore check if ancestor has a name. If it does not,
it's probably safe to cut off any speech associated with it.