Remove __addVerbalizedPunctuation
In Orca v45, we turned over the job to present characters, including
punctuation, over to speech-dispatcher and the synthesizers. But we
took a just-in-case approach and stopped using the chnames dict. So
far so good. Therefore, start removing the associated logic.

The only substitution still relevant in __addVerbalizedPunctuation
was that of subscripted/superscripted unicode numbers. To preserve
that, move it to adjustForPronunciation in script_utilities.py.

See issue #313.