Stop using the chnames dictionary in keynames.py
In Orca v45, we turned over the job to present characters, including
punctuation, over to speech-dispatcher and the synthesizers. But we
took a just-in-case approach and stopped using the chnames dict. So
far so good. Therefore, start removing the associated logic.

In getKeyName, chnames.getCharacterName(key) returns key so fall back
on key directly.

See issue #313.