Install python tools in a venv, not on the system
openSUSE has moved to Python311, whose pip doesn't like to install
things in the system.  Instead, it prefers us to create a virtual
environment and put stuff there.

So, we create /usr/local/python as a venv and install ruff there.

We can put python stuff in ci/requirements.txt and it will get
installed in that venv.  It only has "ruff" now, but we can add other
tools in the future (Sphinx, pytest, etc.).