Meson: Use python to check Gtk dependency met
Checking for Gtk via dependency() can fail if the end user does not
have the Gtk development libraries installed. Therefore do this
check via python. It's messier, but less likely to impact end users
testing Orca.

While we could do something similar for the other dependencies, end
user testing so far suggests this is not needed. And dependency()
is so much cleaner...