Remove no-longer-needed script from caret_navigation.py
All of the handlers/callbacks include a script argument, so we don't
need to keep track of the script during initialization any more.

In addition, some TBD issue is causing the app associated with the
script to get set to None. Under that circumstance, we were failing
to (re) set up keybindings and grabs. While that needs to be solved
at the source of the problem, this commit should cause us to set
up keybindings and grabs when the caller asks us to.