Adjust position-based check when finding on-screen objects
Now that we're using window coordinates instead of screen coordinates,
we're sometimes getting positions that are negative for things which
are fully on-screen. Until we track down the source of that issue,
adjust the position check to take the extents into account before
concluding a container is not on screen.