Mouse Review: Use coordinates from event and a queue
We were using Gdk.Device.get_position in order to determine if the
mouse pointer moved again so we could return early. Instead of doing
this, queue events and return early if the event we're about to
process is not the most recent one. Because the mouse:abs event
includes the coordinates, there's no need to get the position via
Gdk.Device.