Remove two obsolete functions from Gecko script
The default _treatAsLeafNode handles the same condition as the Gecko
script.

The isOnScreen override was to work around an old Gecko bug in which
objects with an "unknown" role were being included as on-screen. That
bug is presumed fixed. And if it still exists, let's get it fixed in
in Gecko.