Mouse Review: Check active state when there are multiple candidate windows
In certain circumstances we can have multiple candidate accessible
windows under the pointer, all of which have a rect which contains
the pointer and the same accessible name. When this occurs, we would
prefer the candidate whose rect (from libwnck) matches the rect reported
to us via AT-SPI2. Those rects are not guaranteed to match for reasons
yet to be determined. Regardless, if we have more than one matching
candidate, the active window is likely on top and contains the object
we want to present. Therefore, start using the accessible state and
stop comparing rects.