Remove isZombie; use AXObject.is_valid instead
isZombie did a bunch of checks and was put in place to handle various and
sundry brokenness in applications and toolkits. Objects that we really
should not touch/present should not pass the AXObject.is_valid test. If
there are other objects which pass the AXObject.is_valid test, but have
other brokenness, let's get the bug fixed and stop doing non-performant
workarounds.